Move the indexing subscribers to the subscriber file
