+++ h/app.py
@@ -114,7 +114,6 @@ def includeme(config):
     config.include("h.db")
     config.include("h.eventqueue")
     config.include("h.form")
-    config.include("h.indexer")
     config.include("h.panels")
     config.include("h.realtime")
     config.include("h.renderers")
